+++ opal/hospital_settings/tests/test_models.py
@@ -1,8 +1,11 @@
-from django.db import models
+from django.core.exceptions import ValidationError
+from django.db import IntegrityError, models
 
 import pytest
+from pytest_django.asserts import assertRaisesMessage
 
-from ..models import Institution, Site
+from .. import constants, factories
+from ..models import Institution, RelationshipType, Site
 
 TEST_HOSPITAL = 'Test Hospital'
 
@@ -304,3 +307,73 @@ def test_site_ordered() -> None:
     Site.objects.create(name='ATest Site', code='TST2', institution=institution)
     first = Site.objects.all()[0]
     assert first.name == 'ATest Site'
+
+
+def test_relationshiptype_str() -> None:
+    """Ensure the `__str__` method is defined for the `RelationshipType` model."""
+    relationship_type = RelationshipType(name='Test User Patient Relationship Type')
+    assert str(relationship_type) == 'Test User Patient Relationship Type'
+
+
+def test_relationshiptype_duplicate_names() -> None:
+    """Ensure that the relationship type name is unique."""
+    factories.RelationshipType(name='Self')
+
+    with assertRaisesMessage(IntegrityError, "Duplicate entry 'Self' for key 'name'"):  # type: ignore[arg-type]
+        factories.RelationshipType(name='Self')
+
+
+def test_relationshiptype_min_age_lowerbound() -> None:
+    """Ensure the minimum age lower bound is validated correctly."""
+    relationship_type = factories.RelationshipType()
+    relationship_type.start_age = constants.RELATIONSHIP_MIN_AGE - 1
+
+    message = 'Ensure this value is greater than or equal to {0}.'.format(constants.RELATIONSHIP_MIN_AGE)
+
+    with assertRaisesMessage(ValidationError, message):  # type: ignore[arg-type]
+        relationship_type.full_clean()
+
+    relationship_type.start_age = constants.RELATIONSHIP_MIN_AGE
+    relationship_type.full_clean()
+
+
+def test_relationshiptype_min_age_upperbound() -> None:
+    """Ensure the minimum age upper bound is validated correctly."""
+    relationship_type = factories.RelationshipType()
+    relationship_type.start_age = constants.RELATIONSHIP_MAX_AGE
+
+    message = 'Ensure this value is less than or equal to {0}.'.format(constants.RELATIONSHIP_MAX_AGE - 1)
+
+    with assertRaisesMessage(ValidationError, message):  # type: ignore[arg-type]
+        relationship_type.full_clean()
+
+    relationship_type.start_age = constants.RELATIONSHIP_MAX_AGE - 1
+    relationship_type.full_clean()
+
+
+def test_relationshiptype_max_age_lowerbound() -> None:
+    """Ensure the maximum age lower bound is validated correctly."""
+    relationship_type = factories.RelationshipType()
+    relationship_type.end_age = constants.RELATIONSHIP_MIN_AGE
+
+    message = 'Ensure this value is greater than or equal to {0}.'.format(constants.RELATIONSHIP_MIN_AGE + 1)
+
+    with assertRaisesMessage(ValidationError, message):  # type: ignore[arg-type]
+        relationship_type.full_clean()
+
+    relationship_type.end_age = constants.RELATIONSHIP_MIN_AGE + 1
+    relationship_type.full_clean()
+
+
+def test_relationshiptype_max_age_upperbound() -> None:
+    """Ensure the maximum age upper bound is validated correctly."""
+    relationship_type = factories.RelationshipType()
+    relationship_type.end_age = constants.RELATIONSHIP_MAX_AGE + 1
+
+    message = 'Ensure this value is less than or equal to {0}.'.format(constants.RELATIONSHIP_MAX_AGE)
+
+    with assertRaisesMessage(ValidationError, message):  # type: ignore[arg-type]
+        relationship_type.full_clean()
+
+    relationship_type.end_age = constants.RELATIONSHIP_MAX_AGE
+    relationship_type.full_clean()