+++ CHANGELOG.md
@@ -20,6 +20,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 * Users app with custom user model (!39)
 * Enforce authenticated requests to the API endpoints (!41)
 * Enforce authenticated user for all URLs (!72)
+* feat: Add direction url field for sites (!73)
+* feat: Add user patient relationship type model (!84)
 
 ### Documentation
 
@@ -38,4 +40,3 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 * chore: Add vscode extension recommendations (!40)
 * build: Fix broken documentation build (!52)
 * i18n: Add French translations for the hospital settings app (!46)
-* feat: Add direction url field for sites